web/requirements.txt
Django==1.11
djangorestframework==3.6.2
gunicorn==19.7.1
psycopg2==2.7.1
redis==2.10.5

django-debug-toolbar
django-widget-tweaks
django-haystack
django-oscar
django-oscar-api
django-redis
django-environ

whoosh==2.7.0
mistune
Pillow==3.1.0


elasticsearch>=5,<6
certifi

wagtail

markdown
django-filter
django-cors-headers

django-taggit
django-allauth

pycountry
better-exceptions
django_extensions
django_compressor
whitenoise